Update to assets server
