+++ packages/test/src/cli.ts
@@ -6,10 +6,10 @@ import { tsImport } from 'tsx/esm/api'
 import { runBrowserTests } from './lib/runner-browser.ts'
 import { runServerTests } from './lib/runner.ts'
 import { createReporter } from './lib/reporters/index.ts'
+import type { Counts } from './lib/reporters/results.ts'
 import { createWatcher } from './lib/watcher.ts'
 import { loadPlaywrightConfig, resolveProjects } from './lib/playwright.ts'
 import { IS_RUNNING_FROM_SRC, loadConfig, type ResolvedRemixTestConfig } from './lib/config.ts'
-import type { Counts } from './lib/utils.ts'
 
 const config = await loadConfig()
