+++ packages/test/src/lib/reporters/dot.ts
@@ -1,6 +1,7 @@
-import { colors, normalizeLine, type Counts } from '../utils.ts'
-import type { TestResult, TestResults } from '../executor.ts'
+import { colors } from '../colors.ts'
+import { normalizeLine } from '../normalize.ts'
 import type { Reporter } from './index.ts'
+import type { Counts, TestResult, TestResults } from './results.ts'
 
 export class DotReporter implements Reporter {
   #failures: { name: string; error: TestResult['error'] }[] = []
