+++ packages/test/src/lib/reporters/tap.ts
@@ -1,6 +1,6 @@
-import { normalizeLine, type Counts } from '../utils.ts'
-import type { TestResults } from '../executor.ts'
 import type { Reporter } from './index.ts'
+import type { TestResults, Counts } from './results.ts'
+import { normalizeLine } from '../normalize.ts'
 
 export class TapReporter implements Reporter {
   #counter = 0
