+++ packages/test/src/lib/worker-e2e.ts
@@ -1,12 +1,13 @@
 import { workerData, parentPort } from 'node:worker_threads'
 import { tsImport } from 'tsx/esm/api'
 import { createServer } from './e2e-server.ts'
-import { runTests, type TestResults } from './executor.ts'
+import { runTests } from './executor.ts'
 import {
   getBrowserLauncher,
   getPlaywrightLaunchOptions,
   getPlaywrightPageOptions,
 } from './playwright.ts'
+import type { TestResults } from './reporters/results.ts'
 
 try {
   await tsImport(workerData.file, import.meta.url)
